Move ThreadRunner into threads module
